+++ src/main/java/com/trybe/acc/java/minhasseries/repository/SerieRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SerieRepository extends JpaRepository<Serie, Long> {
+public interface SerieRepository extends JpaRepository<Serie, Integer> {
   boolean existsByNome(String nome);
 }
